processor_service/requirements.txt
kafka-python==2.0.2
elasticsearch==7.13.4
